api/urls.py
from django.urls import path

from . import views
from rest_framework.routers import DefaultRouter

router = DefaultRouter()
router.register(r'users', views.UserViewSet, basename='user')

urlpatterns = [
    path("driver-location/", views.index, name="Driver update location"),
] + router.urls
