requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2.1
django-eventstream==4.5.1
django-filter==23.2
django-grip==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gripcontrol==4.1.0
idna==3.4
Markdown==3.4.3
MarkupSafe==2.1.2
mysqlclient==2.1.1
pubcontrol==3.4.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==0.9.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.3.4
